config: allow to build tarball when yacc is missing
Autoconf does not handle this scenario as well as Automake does.
If gram.c is already there, do not abort the configuration
if yacc is missing as it is not needed.

If both gram.c and yacc are missing, abort configuration.
Yacc is a required tool to build the package.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>