Fix SEGV if CreateColormapWindow() returns NULL.
twm can segfault when it cannot collect attribute information on
a window that has been destroyed before it got a chance to collect
it's attributes.
twm seems to assume that it is always able to collect this information
and disregards the error if the function that attempts to collect it
fails and CreateColormapWindow() returns NULL.
If this happens correct number_cmap_windows to prevent a SEGV later
on due to an incorrect setting.

Signed-off-by: Egbert Eich <eich@freedesktop.org>