Remove custom version of system(3c) for Solaris
Was originally added in 1993 with the comment:
 make f.exec do the right thing to get working xterms under Solaris2.

Doesn't seem to be needed any more on modern Solaris.
As a bonus, gets rid of error found by Oracle Parfait 1.5.1 that
if the execl() call fails, the function falls through and returns
an uninitialized value for status and then both parent & child
go back to running the same code.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>